refactored the functions to use the functionbase
