+++ JLio.Core/Models/CommandBase.cs
@@ -1,4 +1,5 @@
 ﻿using JLio.Core.Contracts;
+using JLio.Core.Extensions;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
@@ -7,18 +8,10 @@ namespace JLio.Core.Models
     public abstract class CommandBase : ICommand
     {
         [JsonProperty("command")]
-        public string CommandName => FirstCharToLowerCase(GetType().Name);
+        public string CommandName => GetType().Name.CamelCasing();
 
         public abstract JLioExecutionResult Execute(JToken dataContext, IExecutionOptions options);
 
         public abstract ValidationResult ValidateCommandInstance();
-
-        private string FirstCharToLowerCase(string str)
-        {
-            if (string.IsNullOrEmpty(str) || char.IsLower(str[0]))
-                return str;
-
-            return char.ToLower(str[0]) + str.Substring(1);
-        }
     }
 }
