+++ JLio.Functions/NewGuid.cs
@@ -7,10 +7,6 @@ namespace JLio.Functions
 {
     public class NewGuid : FunctionBase
     {
-        public NewGuid() : base("newGuid")
-        {
-        }
-
         public override JLioExecutionResult Execute(JToken currentToken, JToken dataContext, IExecutionOptions options)
         {
             return new JLioExecutionResult(true, new JValue(Guid.NewGuid().ToString()));
